requirements.txt
absl-py==1.3.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
bleach==5.0.1
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clip==1.0
cloudpickle==2.2.0
colorama==0.4.5
colorcet==3.0.1
comm==0.1.2
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.28
dask==2022.10.0
datashader==0.14.2
datashape==0.5.2
debugpy==1.6.4
decorator==5.1.1
distributed==2022.10.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.38.0
fsspec==2022.10.0
ftfy==6.1.1
gast==0.4.0
gensim==4.2.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
HeapDict==1.0.1
holoviews==1.15.1
huggingface-hub==0.10.1
idna==3.4
imageio==2.22.2
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jupyter_client==7.4.8
jupyter_core==5.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
msgpack==1.0.4
multipledispatch==0.6.0
nest-asyncio==1.5.6
networkx==2.8.7
nltk==3.7
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
panel==0.14.0
param==1.12.2
parso==0.8.3
partd==1.3.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyct==0.4.8
Pygments==2.13.0
pynndescent==0.5.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pyviz-comms==2.2.1
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==6.2.0
sortedcontainers==2.4.0
stack-data==0.6.2
tblib==1.7.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.13.1
toolz==0.12.0
torch==1.12.1+cu113
torch-cluster==1.6.0+pt112cu113
torch-geometric==2.2.0
torch-scatter==2.1.0+pt112cu113
torch-sparse==0.6.15+pt112cu113
torch-spline-conv==1.2.1+pt112cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.7.0
transformers==4.23.1
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
xarray==2022.10.0
zict==2.2.0
